[Keychron K8 RGB] Update keychron k8 keymaps, and add via json file for k8 iso. (#164)
* Keychron k8 cleanup for via

* Update device descriptors for k8

* update k8 descriptions

* Rearrange files, and add readme

* Match keychron original iso keymap

* F4 is explorer open on keychron layout
