+++ keyboards/keychron/k8/k8_iso.json
@@ -1,0 +1,17 @@
+{
+  "name": "Keychron K8",
+  "vendorId": "0x3434",
+  "productId": "0xfe2f",
+  "lighting": "none",
+  "matrix": {"rows": 6, "cols": 17},
+  "layouts": {
+    "keymap": [
+      ["0,0", {"x": 1}, "0,1", "0,2", "0,3", "0,4", {"x": 0.5}, "0,5", "0,6", "0,7", "0,8", {"x": 0.5}, "0,9", "0,10", "0,11", "0,12", {"x": 0.25}, "0,14", "0,15", "0,16"],
+      [{"y": 0.25}, "1,0", "1,1", "1,2", "1,3", "1,4", "1,5", "1,6", "1,7", "1,8", "1,9", "1,10", "1,11", "1,12", {"w": 2}, "1,13", {"x": 0.25 }, "1,14", "1,15", "1,16"],
+      [{"w": 1.5 }, "2,0", "2,1", "2,2", "2,3", "2,4", "2,5", "2,6", "2,7", "2,8", "2,9", "2,10", "2,11", "2,12", {"x":0.25,"w":1.25,"h":2,"w2":1.5,"h2":1,"x2":-0.25}, "3,13", {   "x": 0.25 }, "2,14", "2,15", "2,16"],
+      [{"w": 1.75 }, "3,0", "3,1", "3,2", "3,3", "3,4", "3,5", "3,6", "3,7", "3,8", "3,9", "3,10", "3,11", "3,12"],
+      [{"w": 1.25 }, "4,0", "4,1", "4,2", "4,3", "4,4", "4,5", "4,6", "4,7", "4,8", "4,9", "4,10", "4,11", {"w": 2.75 }, "4,13", {"x": 1.25 }, "4,15"],
+      [{"w": 1.25 }, "5,0", {"w": 1.25 }, "5,1", {"w": 1.25 }, "5,2", {"w": 6.25 }, "5,6", {"w": 1.25 }, "5,10", {"w": 1.25 }, "5,11", {"w": 1.25 }, "5,12", {"w": 1.25 }, "5,13", {"x": 0.25 }, "5,14", "5,15", "5,16"]
+    ]
+  }
+}
