+++ keyboards/keychron/k8/keymaps/iso/keymap.c
@@ -1,0 +1,97 @@
+/*
+Copyright 2020 Dimitris Mantzouranis
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 2 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+#include QMK_KEYBOARD_H
+
+const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
+
+
+/*				+--------------------------------------------------------------------------+----------------+
+				| ESC |  | F1 | F2 | F3 | F4 | | F5 | F6 | F7 | F8 | | F9| F10| F11| F12|  | |PSCR|????|PAUS|
+				+--------------------------------------------------------------------------+------|----|----|
+ 				|  ~  |  1 |  2 |  3 |  4 |  5 |  6 |  7 |  8 |  9 |  0 |  - |  = | BACKSP | |INS |SCRL|PGUP|
+ 				+--------------------------------------------------------------------------+------|----|----|
+ 				|  TAB  |  Q |  W |  E |  R |  T |  Y |  U |  I |  O |  P |  [ |  ] |      | |DEL |END |PGDN|
+ 				+--------------------------------------------------------------------  RET |------|----|----|
+ 				| CAPSLCK  |  A |  S |  D |  F |  G |  H |  J |  K |  L | ; | ' |  # |     |                |
+ 				+--------------------------------------------------------------------------+      |----|    |
+ 				| LSHIFT | \ |  Z |  X |  C |  V |  B |  N |  M | , | . |  / |   RSHIFT    |      | UP |    |
+ 				+--------------------------------------------------------------------------+------|----|----|
+ 				|LCTRL| LGUI| LALT |            SPACE            | RALT| RGUI | FN | RCTRL | |LFT |DWN |RGT |
+ 				+--------------------------------------------------------------------------+----------------+
+*/
+    /*  Row:       0           1          2          3        4        5        6         7        8        9          10         11         12         13         14         15         16    */
+    [0] = {     {  KC_ESC,     KC_F1,     KC_F2,     KC_F3,   KC_F4,   KC_F5,   KC_F6,    KC_F7,   KC_F8,   KC_F9,     KC_F10,    KC_F11,    KC_F12,    KC_NO,     KC_PSCR,   KC_SLCK,   KC__MUTE},
+                {  KC_GRV,     KC_1,      KC_2,      KC_3,    KC_4,    KC_5,    KC_6,     KC_7,    KC_8,    KC_9,      KC_0,      KC_MINS,   KC_EQL,    KC_BSPC,   KC_INS,    KC_HOME,   KC_PGUP},
+                {  KC_TAB,     KC_Q,      KC_W,      KC_E,    KC_R,    KC_T,    KC_Y,     KC_U,    KC_I,    KC_O,      KC_P,      KC_LBRC,   KC_RBRC,   KC_BSLS,   KC_DEL,    KC_END,    KC_PGDN},
+                {  KC_CAPS,    KC_A,      KC_S,      KC_D,    KC_F,    KC_G,    KC_H,     KC_J,    KC_K,    KC_L,      KC_SCLN,   KC_QUOT,   KC_NUHS,   KC_ENT,    KC_NO,     KC_NO,     KC_NO},
+                {  KC_LSFT,    KC_NUBS,   KC_Z,      KC_X,    KC_C,    KC_V,    KC_B,     KC_N,    KC_M,    KC_COMM,   KC_DOT,    KC_SLSH,   KC_NO,     KC_RSFT,   KC_NO,     KC_UP,     KC_NO},
+                {  KC_LCTL,    KC_LGUI,   KC_LALT,   KC_NO,   KC_NO,   KC_NO,   KC_SPC,   KC_NO,   KC_NO,   KC_NO,     KC_RALT,   KC_RGUI,   MO(1),     KC_RCTL,   KC_LEFT,   KC_DOWN,   KC_RGHT}
+              },
+
+    /*  Row:       0        1                   2                 3        4               5        6        7        8        9        10        11       12         13         14                15        16        */
+    [1] =   {   {    RESET, KC_BRIGHTNESS_DOWN, KC_BRIGHTNESS_UP, KC_APP,  KC_MY_COMPUTER, RGB_VAD, RGB_VAI, KC_MPRV, KC_MPLY, KC_MNXT, KC__MUTE, KC_VOLU, KC_VOLD, KC_NO,   KC_PSCR,          KC_SLEP, RGB_TOG },
+                {  _______, _______,            _______,          _______, _______,        _______, _______, _______, _______, _______, _______,  _______, _______, _______, RGB_SAI,          RGB_SAI, RGB_HUI },
+                {  _______, _______,            _______,          _______, _______,        _______, _______, _______, _______, _______, _______,  _______, _______, _______, RGB_SAD,          RGB_SAD, RGB_HUD },
+                {  _______, _______,            _______,          _______, _______,        _______, _______, _______, _______, _______, _______,  _______, _______, _______, KC_NO,            KC_NO,   KC_NO },
+                {  _______, _______,            _______,          _______, _______,        _______, _______, _______, _______, _______, _______,  _______, KC_NO,   _______, KC_NO,            KC_VOLU, KC_NO },
+                {  _______, _______,            _______,          KC_NO,   KC_NO,          KC_NO,   _______, KC_NO,   KC_NO,   KC_NO,   _______,  _______, _______, _______, RGB_MODE_REVERSE, KC_VOLD, RGB_MODE_FORWARD }
+              },
+    /*  Row:          0        1        2        3        4        5        6        7        8        9        10       11       12       13       14       15       16    */
+    [2] =   {   {  _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, KC_NO,   _______, _______, _______ },
+                {  _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______ },
+                {  _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______ },
+                {  _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______ },
+                {  _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, KC_NO,   _______, KC_NO,   _______, KC_NO   },
+                {  _______, _______, _______, KC_NO,   KC_NO,   KC_NO,   _______, KC_NO,   KC_NO,   KC_NO,   _______, _______, _______, _______, _______, _______, _______ }
+            },
+
+    /*  Row:          0        1        2        3        4        5        6        7        8        9        10       11       12       13       14       15       16    */
+    [3] =   {   {  _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, KC_NO,   _______, _______, _______ },
+                {  _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______ },
+                {  _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______ },
+                {  _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______ },
+                {  _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, KC_NO,   _______, KC_NO,   _______, KC_NO   },
+                {  _______, _______, _______, KC_NO,   KC_NO,   KC_NO,   _______, KC_NO,   KC_NO,   KC_NO,   _______, _______, _______, _______, _______, _______, _______ }
+    }
+};
+void dip_switch_update_user(uint8_t index, bool active){
+  switch(index){
+    case 0:
+      if(active){ //BT mode
+// do stuff
+      }
+      else{ //Cable mode
+// do stuff
+      }
+      break;
+    case 1:
+      if(active){ // Win/Android mode
+// do stuff
+      }
+      else{ // Mac/iOS mode
+// do stuff
+      }
+      break;
+  }
+}
+void keyboard_post_init_user(void) {
+  // Customise these values to desired behaviour
+  // debug_enable=true;
+  // debug_matrix=true;
+  // debug_keyboard=true;
+  // debug_mouse=true;
+}
