+++ keyboards/keychron/k8/keymaps/iso/rules.mk
@@ -1,0 +1,14 @@
+# https://beta.docs.qmk.fm/developing-qmk/qmk-reference/getting_started_make_guide#rules-mk-options
+
+# AUTO_SHIFT_ENABLE = yes
+# MOUSEKEY_ENABLE = yes
+EXTRAKEY_ENABLE = yes    # Audio control and System control
+# MAGIC_ENABLE = yes
+# MAGIC_KEYCODE_ENABLE = yes
+
+# BOOTMAGIC_ENABLE = no	# https://beta.docs.qmk.fm/using-qmk/hardware-features/feature_bootmagic
+# NKRO_ENABLE = no
+
+
+
+
