+++ keyboards/keychron/k8/via/rules.mk
@@ -1,0 +1,5 @@
+VIA_ENABLE = yes
+
+
+
+
